fix: Correct conditional in Publish to PyPI Release workflow job (#137)
The conditional for the Publish to PyPI job incorrectly checked the
`ref_type`. It's done correctly in the Validate tag job. This didn't
cause any issues since the Publish to PyPI job requires the Validate tag
job, but this fix should prevent any future issues if the dependency is
removed.