Compute pair cub features (#162)
**Summary of this PR:**

- added variables to the Pairs class to store gene-level features
- updated do_setup to generate codon and amino acid profiles for each
organism
- created CodonBiasComparison objects in compute_feature() method for
each type of codon/aa profile, to compare a single virus and single host
- compute comparisons (slope, R2, cosine similarity) on each
CodonBiasComparison object (for a single pair) using built-in methods
- test that all new gene-level pair properties are populated correctly
when compute_feature() is run


**Planning next steps (future PRs):**

- *above, 'gene-level features' = gene-level features based on sequence
only (i.e. variations of CUB)...next up is extracting features
considering function!
- there were several pre-existing '# pyright: ignore' statements in the
compute_feature() definition, most avoiding unchecked None types - I
would like to clarify this and remove ignore statements
- _TO DISCUSS:_ pairs_of_interest / determine_custom_pairs could have
been useful for testing here (i.e. we don't need to calculate everything
for all pairs if we are only interested in a certain pair), but maybe
this concept is not yet well-integrated in the rest of the code? For
example, generate_kmer_profiles only works on ALL files given in the
genome directories - there is no option to just compute these profiles
for a single pair of interest. It would help to know what the original
intention for this option was? As it stands, determine_custom_pairs
appends pairs to the existing pairs list...are pairs of interest meant
for adding any files outside the main directories? Or is this meant to
focus all the calculations to one or a few pairs?
- one last thing I noticed: inconsistency in using/not using os to
concatenate file paths...all work, but worth cleaning up

---------

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: Eric Bastien <gbastien@umich.edu>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>